Update dependabot.yml
We've configured Dependabot to automatic dependency updates so that our project, which is built on Vue.js and Node.js, stays safe and current. This setup keeps track of the dependencies for our project that are specified in the package.json file and looks for any updates or security flaws.